refactor(gwr-components,gwr-models): separate benchmark harness and simulation functions
