+++ gwr-components/benches/components.rs
@@ -2,9 +2,6 @@
 
 //! Benchmark basic component usage.
 
-use std::time::Duration;
-
-use criterion::{BatchSize, Criterion, criterion_group};
 use gwr_components::arbiter::Arbiter;
 use gwr_components::arbiter::policy::{Priority, RoundRobin};
 use gwr_components::delay::Delay;
@@ -24,11 +21,12 @@ fn create_engine() -> Engine {
     Engine::new(&tracker)
 }
 
-fn run_engine(mut engine: Engine) {
+pub fn run_engine(mut engine: Engine) {
     engine.run().unwrap();
 }
 
-fn spawn_source_store_sink() -> Engine {
+#[must_use]
+pub fn spawn_source_store_sink() -> Engine {
     let mut engine = create_engine();
     let clock = engine.default_clock();
 
@@ -47,7 +45,8 @@ fn spawn_source_store_sink() -> Engine {
     engine
 }
 
-fn spawn_source_delay_sink() -> Engine {
+#[must_use]
+pub fn spawn_source_delay_sink() -> Engine {
     let mut engine = create_engine();
     let clock = engine.default_clock();
 
@@ -67,7 +66,8 @@ fn spawn_source_delay_sink() -> Engine {
     engine
 }
 
-fn spawn_arbiter_fixedpriority_policy() -> Engine {
+#[must_use]
+pub fn spawn_arbiter_fixedpriority_policy() -> Engine {
     let mut engine = create_engine();
     let inputs = vec![
         ArbiterInputData {
@@ -100,7 +100,8 @@ fn spawn_arbiter_fixedpriority_policy() -> Engine {
     engine
 }
 
-fn spawn_larger_simulation() -> Engine {
+#[must_use]
+pub fn spawn_larger_simulation() -> Engine {
     let mut engine = create_engine();
     let clock = engine.default_clock();
     let rate_limiter = rc_limiter!(&clock, 1);
@@ -140,56 +141,4 @@ fn spawn_larger_simulation() -> Engine {
     connect_port!(sink_limiter, tx => sink, rx).unwrap();
 
     engine
-}
-
-fn bench_small_blocks(c: &mut Criterion) {
-    let mut group = c.benchmark_group("blocks");
-
-    group.bench_function("source_store_sink", |b| {
-        b.iter_batched(spawn_source_store_sink, run_engine, BatchSize::SmallInput);
-    });
-
-    group.bench_function("source_delay_sink", |b| {
-        b.iter_batched(spawn_source_delay_sink, run_engine, BatchSize::SmallInput);
-    });
-
-    group.bench_function("arbiter_fixedpriority_policy", |b| {
-        b.iter_batched(
-            spawn_arbiter_fixedpriority_policy,
-            run_engine,
-            BatchSize::SmallInput,
-        );
-    });
-
-    group.finish();
-}
-
-fn bench_large_blocks(c: &mut Criterion) {
-    let mut group = c.benchmark_group("blocks");
-
-    group.bench_function("larger_simulation", |b| {
-        b.iter_batched(spawn_larger_simulation, run_engine, BatchSize::SmallInput);
-    });
-
-    group.finish();
-}
-
-criterion_group! {
-    name = small_benches;
-    config = Criterion::default()
-        .measurement_time(Duration::from_secs(10))
-        .warm_up_time(Duration::from_secs(6))
-        .noise_threshold(0.05)
-        .confidence_level(0.98)
-        .significance_level(0.03);
-    targets = bench_small_blocks
-}
-
-criterion_group! {
-    name = large_benches;
-    config = Criterion::default()
-        .measurement_time(Duration::from_secs(10))
-        .noise_threshold(0.02)
-        .confidence_level(0.98);
-    targets = bench_large_blocks
 }
