+++ gwr-models/benches/ethernet_frame.rs
@@ -1,9 +1,7 @@
 // Copyright (c) 2025 Graphcore Ltd. All rights reserved.
 
 use std::rc::Rc;
-use std::time::Duration;
 
-use criterion::{BatchSize, Criterion, criterion_group};
 use gwr_components::connect_port;
 use gwr_components::sink::Sink;
 use gwr_components::store::Store;
@@ -22,7 +20,7 @@ fn create_engine() -> Engine {
     Engine::new(&tracker)
 }
 
-fn run_engine<T>(args: (Engine, Rc<Sink<T>>))
+pub fn run_engine<T>(args: (Engine, Rc<Sink<T>>))
 where
     T: SimObject,
 {
@@ -31,7 +29,8 @@ where
     assert_eq!(sink.num_sunk(), NUM_FRAMES);
 }
 
-fn setup_frame_simulation() -> (Engine, Rc<Sink<EthernetFrame>>) {
+#[must_use]
+pub fn setup_frame_simulation() -> (Engine, Rc<Sink<EthernetFrame>>) {
     let num_frames = NUM_FRAMES;
     let payload_size_bytes = 256;
     let frame_dest = [0, 1, 2, 3, 4, 5];
@@ -68,7 +67,8 @@ fn setup_frame_simulation() -> (Engine, Rc<Sink<EthernetFrame>>) {
     (engine, sink)
 }
 
-fn setup_box_frame_simulation() -> (Engine, Rc<Sink<Box<EthernetFrame>>>) {
+#[must_use]
+pub fn setup_box_frame_simulation() -> (Engine, Rc<Sink<Box<EthernetFrame>>>) {
     let num_frames = NUM_FRAMES;
     let payload_size_bytes = 256;
     let frame_dest = [0, 1, 2, 3, 4, 5];
@@ -103,32 +103,4 @@ fn setup_box_frame_simulation() -> (Engine, Rc<Sink<Box<EthernetFrame>>>) {
     connect_port!(store, tx => sink, rx).unwrap();
 
     (engine, sink)
-}
-
-fn bench_ethernet_frame(c: &mut Criterion) {
-    let mut group = c.benchmark_group("ethernet_frame");
-
-    group.bench_function("vec_of_frame", |b| {
-        b.iter_batched(setup_frame_simulation, run_engine, BatchSize::SmallInput);
-    });
-
-    group.bench_function("vec_of_box", |b| {
-        b.iter_batched(
-            setup_box_frame_simulation,
-            run_engine,
-            BatchSize::SmallInput,
-        );
-    });
-
-    group.finish();
-}
-
-criterion_group! {
-    name = benches;
-    config = Criterion::default()
-        .measurement_time(Duration::from_secs(10))
-        .warm_up_time(Duration::from_secs(6))
-        .noise_threshold(0.03)
-        .confidence_level(0.98);
-    targets = bench_ethernet_frame
 }
